requirements.txt
PyQt6
requests
pyqtgraph
python-dotenv
